--- team3/front-end/src/app/pages/LoginPage.jsx
@@ -1,9 +1,0 @@
-import { Login } from "../components/pages/Login";
-
-export const LoginPage = () => {
-  return (
-    <div>
-      <Login />
-    </div>
-  );
-};
