--- team3/front-end/src/app/pages/SignUpPage.jsx
@@ -1,9 +1,0 @@
-import { SignUp } from "../components/pages/SignUp";
-
-export const SignUpPage = () => {
-  return (
-    <div>
-      <SignUp />
-    </div>
-  );
-};
