+++ team3/front-end/src/app/sign-up/page.jsx
@@ -1,0 +1,9 @@
+import { SignUpPage } from "../components/pages/SignUpPage";
+
+export default function Page() {
+  return (
+    <div>
+      <SignUpPage />
+    </div>
+  );
+}
